@@ -1,7 +1,7 @@
 package jonasbark.de.stripepaymentexample
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import de.jonasbark.stripepayment.StripeDialog
 
 class MainActivity : AppCompatActivity() {
